Revert "rebase: allow starting from a dirty tree."
This reverts commit 48dd40fef4208944b7ba27ee7b8507f0073a7598.
Allowing rebase to start in a dirty tree might have been a worthy
goal, but it is not necessarily always wanted (some people prefer
to be reminded that the state is dirty, and think about the next
action that may not be to stash and proceed).  Furthermore, depending
on the nature of local changes, unstashing the dirty state on top of
the rebased result is not always desirable.

Signed-off-by: Junio C Hamano <gitster@pobox.com>