Revert "builtin-commit.c: remove useless check added by faulty cut and paste"
This reverts commit 0d198e5a7dd59e1f512c1c3bfb57fedc6a6d68a7.

We are calling overlay_tree_on_cache() which does use CE_UPDATE
flag to mark duplicated entries, which is the same as the
codepath in git-ls-files with its --with-tree option.

Because the pathname ce->name is given to path_list_insert()
which does not allow duplicates, there is no breakage either way
from the correctness point of view in this codepath, unlike the
one in ls-files.  But avoiding unnecessary processing with a
single bit check is certainly better.